test: fold Wolfram run summary suite
